seperate lecture and menu
