+++ src/components/LaborLectureMenu/LaborLecture.module.css
