+++ src/components/LaborLectureMenu/LectureEntry.js
